+++ exeris-kernel-community/src/test/java/eu/exeris/kernel/community/bootstrap/CommunityPersistenceBootstrapIntegrationTest.java
@@ -188,7 +188,8 @@ class CommunityPersistenceBootstrapIntegrationTest {
                                             HttpConfig.DEFAULT_MAX_HEADER_SIZE,
                                             HttpConfig.DEFAULT_MAX_REQUEST_BODY_BYTES,
                                             false,
-                                            HttpVersion.HTTP_1_1
+                                            HttpVersion.HTTP_1_1,
+                                            "127.0.0.1" + ":" + port
                                     );
                                     try (HttpClientEngine client = new CommunityHttpProvider().createClientEngine(clientConfig)) {
                                         client.start();